Cleaned up unused scripts, updated utils and main notebook
