--- DATA605/Spring2025/projects/TutorTask137_Spring2025_Bitcoin_Forecasting_with_Facebook_Prophet/scripts/forecast_script.py
@@ -1,12 +1,0 @@
-import matplotlib.pyplot as plt
-
-def make_forecast(model, periods=30):
-    future = model.make_future_dataframe(periods=periods)
-    forecast = model.predict(future)
-    return forecast
-
-def plot_forecast(model, forecast):
-    fig = model.plot(forecast)
-    plt.title("Bitcoin Price Forecast")
-    plt.grid(True)
-    plt.show()
