--- DATA605/Spring2025/projects/TutorTask137_Spring2025_Bitcoin_Forecasting_with_Facebook_Prophet/scripts/prophet_model.py
@@ -1,13 +1,0 @@
-from prophet import Prophet
-
-def create_prophet_model():
-    model = Prophet(
-        yearly_seasonality=True,
-        weekly_seasonality=True,
-        daily_seasonality=False
-    )
-    return model
-
-def train_prophet_model(model, data):
-    model.fit(data)
-    return model
